check file existance

+++ bash_scripts/generacion_tabla_inicio.sh
@@ -36,6 +36,8 @@ while IFS= read -r line; do
   fi
 
 done < $temp # se pone el comando acá en vez de antes del while porque sino se pierde el valor de las variables dentro del bucle
-rm $temp
+if [ ! test -f "$temp" ]; then
+  rm $temp
+fi
 echo "Fin de la consulta de estado"
